Change TogglePanel from control to user control
