--- code/src/UI/Styles/StyleTogglePane.xaml
@@ -1,17 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:controls="clr-namespace:Microsoft.Templates.UI.Controls">
-
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/_Colors.xaml" />
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/_FontSizes.xaml" />
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/_Thickness.xaml" />
-    </ResourceDictionary.MergedDictionaries>
-
-    <Style x:Key="TogglePaneStyle" TargetType="controls:TogglePane">
-        <Setter Property="HorizontalAlignment" Value="Stretch" />
-        <Setter Property="VerticalAlignment" Value="Stretch" />
-        <Setter Property="Focusable" Value="False" />
-    </Style>
-    
-</ResourceDictionary>
