+++ code/src/UI/Styles/UIStyles.xaml
@@ -11,14 +11,13 @@
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleExpander.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleGroupBox.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleHyperlink.xaml" />
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleImage.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleListView.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleLogoDisplay.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StylePanel.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleShape.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleTextBlock.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleTextBox.xaml" />
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleTogglePane.xaml" />
+        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Styles/StyleTogglePanel.xaml" />
     </ResourceDictionary.MergedDictionaries>
     
 </ResourceDictionary>
