+++ code/src/UI/Themes/Generic.xaml
@@ -5,6 +5,5 @@
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Controls/CodeViewer.xaml" />
         <ResourceDictionary Source="/Microsoft.Templates.UI;component/Controls/OverlayBox.xaml" />
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Controls/TogglePane.xaml" />
     </ResourceDictionary.MergedDictionaries>
 </ResourceDictionary>
