+++ code/src/UI/Views/NewProject/MainView.xaml
@@ -39,26 +39,26 @@
             </DataTemplate>
 
             <DataTemplate x:Key="ProjectTemplatesSummaryItemTemplate">
-                <controls:TogglePane TogglePaneContent="{Binding}" IsOpen="{Binding IsOpen}" AllowDragAndDrop="{Binding AllowDragAndDrop}" Style="{StaticResource TogglePaneStyle}" Width="263">
-                    <controls:TogglePane.OpenButtonTemplate>
+                <controls:TogglePanel TogglePanelContent="{Binding}" IsOpen="{Binding IsOpen}" AllowDragAndDrop="{Binding AllowDragAndDrop}" Style="{StaticResource TogglePanelStyle}" Width="263">
+                    <controls:TogglePanel.OpenButtonTemplate>
                         <DataTemplate>
-                            <Button Style="{StaticResource BlueTogglePaneButtonStyle}" Command="{Binding OpenCommand}" CommandParameter="{Binding}" Focusable="True" ToolTip="{x:Static res:StringRes.SummaryOpenMenu}" AutomationProperties.Name="{x:Static res:StringRes.SummaryOpenMenu}">
+                            <Button Style="{StaticResource BlueTogglePanelButtonStyle}" Command="{Binding OpenCommand}" CommandParameter="{Binding}" Focusable="True" ToolTip="{x:Static res:StringRes.SummaryOpenMenu}" AutomationProperties.Name="{x:Static res:StringRes.SummaryOpenMenu}">
                                 <Button.Content>
                                     <TextBlock Text="&#xE713;" FontSize="18" Style="{StaticResource IconButtonTextStyle}"/>
                                 </Button.Content>
                             </Button>
                         </DataTemplate>
-                    </controls:TogglePane.OpenButtonTemplate>
-                    <controls:TogglePane.CloseButtonTemplate>
+                    </controls:TogglePanel.OpenButtonTemplate>
+                    <controls:TogglePanel.CloseButtonTemplate>
                         <DataTemplate>
-                            <Button Style="{StaticResource BlueTogglePaneButtonStyle}" Command="{Binding OpenCommand}" CommandParameter="{Binding}" Focusable="True" ToolTip="{x:Static res:StringRes.SummaryOpenMenuClose}"  AutomationProperties.Name="{x:Static res:StringRes.SummaryOpenMenuClose}">
+                            <Button Style="{StaticResource BlueTogglePanelButtonStyle}" Command="{Binding OpenCommand}" CommandParameter="{Binding}" Focusable="True" ToolTip="{x:Static res:StringRes.SummaryOpenMenuClose}"  AutomationProperties.Name="{x:Static res:StringRes.SummaryOpenMenuClose}">
                                 <Button.Content>
                                     <TextBlock Text="&#xE761;" FontSize="18" Style="{StaticResource IconButtonTextStyle}"/>
                                 </Button.Content>
                             </Button>
                         </DataTemplate>
-                    </controls:TogglePane.CloseButtonTemplate>
-                    <controls:TogglePane.MainViewTemplate>
+                    </controls:TogglePanel.CloseButtonTemplate>
+                    <controls:TogglePanel.MainViewTemplate>
                         <DataTemplate>
                             <Grid Height="30" Focusable="False" Margin="{StaticResource Margin_S_Left}">
                                 <Grid VerticalAlignment="Center">
@@ -72,28 +72,28 @@
                                 </Grid>
                             </Grid>
                         </DataTemplate>
-                    </controls:TogglePane.MainViewTemplate>
-                    <controls:TogglePane.SecondaryViewTemplate>
+                    </controls:TogglePanel.MainViewTemplate>
+                    <controls:TogglePanel.SecondaryViewTemplate>
                         <DataTemplate>
                             <Grid HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Focusable="False" KeyboardNavigation.TabNavigation="Local">
                                 <Grid.ColumnDefinitions>
                                     <ColumnDefinition Width="*" />
                                     <ColumnDefinition Width="*" />
                                 </Grid.ColumnDefinitions>
-                                <Button Grid.Column="0" Style="{StaticResource BlackTogglePaneButtonStyle}" Command="{Binding RenameCommand}" CommandParameter="{Binding}" IsEnabled="{Binding CanChooseItemName}" ToolTip="{x:Static res:StringRes.SummaryOptionRename}" Focusable="{Binding IsOpen}" AutomationProperties.Name="{x:Static res:StringRes.SummaryOptionRename}">
+                                <Button Grid.Column="0" Style="{StaticResource BlackTogglePanelButtonStyle}" Command="{Binding RenameCommand}" CommandParameter="{Binding}" IsEnabled="{Binding CanChooseItemName}" ToolTip="{x:Static res:StringRes.SummaryOptionRename}" Focusable="{Binding IsOpen}" AutomationProperties.Name="{x:Static res:StringRes.SummaryOptionRename}">
                                     <Button.Content>
                                         <TextBlock Text="&#xE70F;" Style="{StaticResource IconButtonTextStyle}" FontSize="16"/>
                                     </Button.Content>
                                 </Button>
-                                <Button Grid.Column="1" Style="{StaticResource BlackTogglePaneButtonStyle}" Command="{Binding RemoveCommand}" CommandParameter="{Binding}" IsEnabled="{Binding CanSetHome}" ToolTip="{x:Static res:StringRes.SummaryOptionDelete}" Focusable="{Binding IsOpen}" AutomationProperties.Name="{x:Static res:StringRes.SummaryOptionDelete}">
+                                <Button Grid.Column="1" Style="{StaticResource BlackTogglePanelButtonStyle}" Command="{Binding RemoveCommand}" CommandParameter="{Binding}" IsEnabled="{Binding CanSetHome}" ToolTip="{x:Static res:StringRes.SummaryOptionDelete}" Focusable="{Binding IsOpen}" AutomationProperties.Name="{x:Static res:StringRes.SummaryOptionDelete}">
                                     <Button.Content>
                                         <TextBlock Text="&#xE74D;" Style="{StaticResource IconButtonTextStyle}" FontSize="15"/>
                                     </Button.Content>
                                 </Button>
                             </Grid>
                         </DataTemplate>
-                    </controls:TogglePane.SecondaryViewTemplate>
-                </controls:TogglePane>
+                    </controls:TogglePanel.SecondaryViewTemplate>
+                </controls:TogglePanel>
             </DataTemplate>
 
             <DataTemplate x:Key="LicenseSummaryItemTemplate">
